+++ src/ramobjectmodels/statisticsmodel.h
@@ -34,7 +34,7 @@ private:
     RamProject *m_project = nullptr;
     RamUser *m_user = nullptr;
 
-    QHash<QString, QVector<float>> m_statsPerStep; // step userstats + step lateness ratio
+    QHash<QString, QVector<qreal>> m_statsPerStep; // step userstats + step lateness ratio
 
     // Connect submodels and relay events
     void connectEvents();
